[ISSUE #10992] Add unit test for SQLParserFactoryTest
